:recycle: Re-work user chip menu using newer patterns from WRAP project.
See https://github.com/icefoganalytics/wrap/blob/d81399f2d61739cb1a7d988a82b8c554deca3e10/web/src/components/users/UserChip.vue
